Made demo data easilly available
