+++ robotremote/RobotRemoteServer.php
@@ -124,37 +124,3 @@ Content-Length: '.$contentLength."\r\n"
 	}
 
 }
-
-class DemoRequests {
-
-	private $requests = array(
-		'<?xml version="1.0"?>
-		<methodCall>
-		   <methodName>get_keyword_names</methodName>
-		   <params>
-		      </params>
-		   </methodCall>',
-		'<?xml version="1.0"?>
-		<methodCall>
-		   <methodName>stop_remote_server</methodName>
-		   <params>
-		      </params>
-		   </methodCall>'
-	);
-	private $requestsIdx = 0;
-
-	public function get() {
-		$request = $this->requests[$this->requestsIdx++];
-		echo('Request: '.$request."\n\n");
-		return $request;
-	}
-
-}
-
-class DemoResponses {
-
-	public function add($response) {
-		echo('Response: '.$response."\n\n");
-	}
-
-}
